+++ types/events_test.go
@@ -92,13 +92,13 @@ func (s *eventsTestSuite) TestEmitTypedEvent() {
 	s.Run("deterministic key-value order", func() {
 		for i := 0; i < 10; i++ {
 			em := sdk.NewEventManager()
-			coin := sdk.NewCoin("fakedenom", math.NewInt(1999999))
+			coin := sdk.NewCoin("fakedenom", sdk.NewInt(1999999))
 			s.Require().NoError(em.EmitTypedEvent(&coin))
 			s.Require().Len(em.Events(), 1)
 			attrs := em.Events()[0].Attributes
 			s.Require().Len(attrs, 2)
-			s.Require().Equal(attrs[0].Key, "amount")
-			s.Require().Equal(attrs[1].Key, "denom")
+			s.Require().Equal(string(attrs[0].Key), "amount")
+			s.Require().Equal(string(attrs[1].Key), "denom")
 		}
 	})
 }